+++ ts/components/Dialog/DialogHeader.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import { useDialogContext } from './DialogContext';
 
 interface DialogHeaderProps {
-  /** Used to set the the id attribute of the dialog heading */
+  /** Used to set the the id attribute of the dialog heading. */
   ariaLabelSelector?: string;
   /** Content for the dialog header. */
   children?: string | JSX.Element[] | JSX.Element;