+++ ts/components/Interstitial.tsx
@@ -2,11 +2,17 @@ import React from 'react';
 import classNames from 'classnames';
 
 interface InterstitialProps {
-  animate?: boolean;
+  /** Content for the interstitial. */
   children: string | JSX.Element[] | JSX.Element;
+  /** Controls if the interstitial should animate in. */
+  animate?: boolean;
+  /** Adds class names to the interstitial. */
   className?: string;
+  /** Adds class names to the interstitial content. */
   contentClassName?: string;
+  /** Adds class names to the interstitial icon. */
   iconClassName?: string;
+  /** Controls the size of the interstitial. */
   size?: `small`;
 }