+++ ts/components/ProgressBar.tsx
@@ -2,8 +2,11 @@ import React from 'react';
 import classNames from 'classnames';
 
 interface ProgressBarProps {
+  /** The number of steps that have been completed. */
   completedSteps: number;
+  /** Set the total number of steps. */
   totalSteps: number;
+  /** Controls the variant of the progress bar. */
   variant?: `default` | `large`;
 }