+++ ts/components/RadioButton.tsx
@@ -4,8 +4,11 @@ import { Label } from './Label';
 import { randomString } from '../helpers/random_string';
 
 interface RadioButtonProps extends InputHTMLAttributes<HTMLInputElement> {
-  label?: string;
+  /** Controls the radio button size. */
   buttonSize?: `small` | `large`;
+  /** Set the label content. */
+  label?: string;
+  /** Controls the radio button variant. */
   variant?: `default` | `pill`;
 }