+++ ts/components/Select/Select.tsx
@@ -25,16 +25,30 @@ interface RenderOpenerProps {
 }
 
 interface SelectProps {
+  /** Content for the dropdown. */
   children: JSX.Element | JSX.Element[];
-  closeOnSelect?: boolean;
-  flip?: boolean;
-  minHeight?: number;
-  maxHeight?: number;
-  height?: string;
-  initialSelectedValue?: string | null;
-  initialSelectedIndex?: number | null;
-  placement?: Placement;
+  /** The dropdown opener element. Use the destructured "ref", "selectedLabel", and "...props" to spread them
+   * onto the opener element for correct functionality. */
   renderOpener: (props: RenderOpenerProps) => JSX.Element;
+  /** Controls whether the dropdown should be closed when an item is selected. */
+  closeOnSelect?: boolean;
+  /** Controls whether the dropdown should flip its orientation based on the available space in the browser
+   * window. */
+  flip?: boolean;
+  /** Controls the height of the dropdown. */
+  height?: string;
+  /** Controls the initial selected value. */
+  initialSelectedValue?: string | null;
+  /** Controls the initial selected index. */
+  initialSelectedIndex?: number | null;
+  /** Controls the maximum height of the dropdown. */
+  maxHeight?: number;
+  /** Controls the minimum height of the dropdown. */
+  minHeight?: number;
+  /** Controls the placement of the dropdown. */
+  placement?: Placement;
+  /** Controls the width of the dropdown. "full" width makes the dropdown the same width as the opener
+   * element. */
   width?: `auto` | `full` | number;
 }