+++ ts/index.ts
@@ -1,3 +1,4 @@
+export { Accordion, AccordionPanelItem } from './components/Accordion';
 export { AnimationDistance } from './types/animation_distance';
 export { AnimationDirection } from './types/animation_direction';
 export { AlertDialog } from './components/Dialog/AlertDialog';