+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.6] - 2024-03-22
+
+### Added
+
+- Added `Accordion` component
+- Added `star-circled--lg` icon
+
+### Updated
+
+- Added missing prop comments to components
+- Recompiled icon font
+
 ## [2.6.5] - 2024-02-14
 
 ### Added